Fix memory size

+++ src/cpu.rs
@@ -48,7 +48,7 @@ impl Cpu {
         match decoded_instruction.opcode {
             OpCode::ADD => self.registers[op0] = self.add(self.registers[op0], self.registers[op1])?,
             OpCode::XOR => self.registers[op0] = self.xor(self.registers[op0], self.registers[op1]),
-            OpCode::LD => self.registers[op0] = self.memory.load(op1)?,
+            OpCode::LD => self.registers[op0] = self.memory.load(op1)? as u16,
             OpCode::ST => self.memory.store(op0, op1 as usize)?
         }
         Ok(())
